requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.2.0
plotly==5.18.0
scipy==1.11.4
